1. Updated migration to digits. 2. Updated controllers to calculate account balance on update of income and expenses. 3. Update currency filters on vue tabs.

+++ database/migrations/2017_04_25_183512_create_accounts_table.php
@@ -19,8 +19,8 @@ class CreateAccountsTable extends Migration
             $table->string('accountName');
             $table->string('accountDescription');
             $table->string('accountType');
-            $table->integer('openingBalance');
-            $table->integer('balance');
+            $table->decimal('openingBalance', 20, 2);
+            $table->decimal('balance', 20, 2);
             $table->softDeletes();
             $table->timestamps();
         });
